+++ packages/plugin-layout/src/utils/getLayoutContent.tsx
@@ -4,11 +4,12 @@ export default (
   userConfig: LayoutConfig,
   path: string,
   formatMessage: boolean,
+  hasAccess: boolean,
 ) => `import React, { useState, useEffect } from "react";
 import { ApplyPluginsType, useModel ${
   // 没有 formatMessage 就不打开国际化
   formatMessage ? `, useIntl` : ''
-}, traverseModifyRoutes, useAccess } from "umi";
+}${hasAccess ? ', traverseModifyRoutes, useAccess' : ''} } from "umi";
 import { plugin } from "../core/umiExports";
 import LayoutComponent from '${path}';
 
@@ -21,7 +22,7 @@ export default props => {
     setInitialState: null
   }; // plugin-initial-state 未开启
 
-  const access = useAccess?.();
+  ${hasAccess ? 'const access = useAccess?.();' : ''}
 
   useEffect(() => {
     const useRuntimeConfig =
@@ -30,9 +31,11 @@ export default props => {
         type: ApplyPluginsType.modify,
         initialValue: {
           ...initialInfo,
-          traverseModifyRoutes: (menuData) => {
-            return traverseModifyRoutes?.(menuData, access);
-          },
+          ${
+            hasAccess
+              ? 'traverseModifyRoutes: (menuData) => {return traverseModifyRoutes?.(menuData, access);},'
+              : ''
+          }
         },
       }) || {};
     if (useRuntimeConfig instanceof Promise) {
@@ -42,7 +45,7 @@ export default props => {
       return;
     }
     setRuntimeConfig(useRuntimeConfig);
-  }, [initialInfo?.initialState, access]);
+  }, [initialInfo?.initialState, ${hasAccess ? 'access' : ''}]);
 
   const userConfig = {
     ...${JSON.stringify(userConfig).replace(/"/g, "'")},
@@ -62,3 +65,88 @@ export default props => {
   });
 };
 `;
+
+const genRenderRightContent = (props: {
+  locale: boolean;
+  initialState: boolean;
+}) => {
+  if (!props.initialState) {
+    return `export default function renderRightContent() {
+    return null;
+  }
+  `;
+  }
+  return `import React from 'react';
+  import { Avatar, Dropdown, Menu, Spin } from 'antd';
+  ${props.locale ? "import { SelectLang } from 'umi';" : ''}
+  import { LogoutOutlined } from '@ant-design/icons';
+  import { ILayoutRuntimeConfig } from '../types/interface.d';
+
+  export default function renderRightContent(
+    runtimeLayout: ILayoutRuntimeConfig,
+    loading: boolean,
+    initialState: any,
+    setInitialState: any,
+  ) {
+    if (runtimeLayout.rightRender) {
+      return runtimeLayout.rightRender(
+        initialState,
+        setInitialState,
+        runtimeLayout,
+      );
+    }
+
+    const menu = (
+      <Menu className="umi-plugin-layout-menu">
+        <Menu.Item
+          key="logout"
+          onClick={() =>
+            runtimeLayout.logout && runtimeLayout?.logout(initialState)
+          }
+        >
+          <LogoutOutlined />
+          退出登录
+        </Menu.Item>
+      </Menu>
+    );
+
+    const avatar = (
+      <span className="umi-plugin-layout-action">
+        <Avatar
+          size="small"
+          className="umi-plugin-layout-avatar"
+          src={
+            initialState?.avatar ||
+            'https://gw.alipayobjects.com/zos/antfincdn/XAosXuNZyF/BiazfanxmamNRoxxVxka.png'
+          }
+          alt="avatar"
+        />
+        <span className="umi-plugin-layout-name">{initialState?.name}</span>
+      </span>
+    );
+
+    if (loading) {
+      return (
+        <div className="umi-plugin-layout-right">
+          <Spin size="small" style={{ marginLeft: 8, marginRight: 8 }} />
+        </div>
+      );
+    }
+
+    return (
+      <div className="umi-plugin-layout-right anticon">
+        {runtimeLayout.logout ? (
+          <Dropdown overlay={menu} overlayClassName="umi-plugin-layout-container">
+            {avatar}
+          </Dropdown>
+        ) : (
+          avatar
+        )}
+        ${props.locale ? '{SelectLang && <SelectLang />}' : ''}
+      </div>
+    );
+  }
+  `;
+};
+
+export { genRenderRightContent };