+++ packages/plugin-qiankun/src/master/modifyRoutes.ts
@@ -4,7 +4,7 @@ import { defaultHistoryType } from '../constants';
 import { App } from '../types';
 
 export default function modifyRoutes(api: IApi) {
-  api.modifyRoutes(routes => {
+  api.modifyRoutes((routes) => {
     const { history, base } = api.config;
     const { master: { routeBindingAlias = 'microApp', apps = [] } = {} } =
       api.config.qiankun || {};
@@ -49,7 +49,9 @@ function modifyRoutesWithAttachMode(
         const { base, masterHistoryType, appName, routeProps } = opts;
         const normalizeJsonStringInUmiRoute = (str: string) =>
           str.replace(/"/g, "'");
-        const normalizedRouteProps = normalizeJsonStringInUmiRoute(JSON.stringify(routeProps));
+        const normalizedRouteProps = normalizeJsonStringInUmiRoute(
+          JSON.stringify(routeProps),
+        );
 
         return `(() => {
           const { getMicroAppRouteComponent } = umiExports;
@@ -57,7 +59,7 @@ function modifyRoutesWithAttachMode(
         })()`;
       };
 
-      routes.forEach(route => {
+      routes.forEach((route) => {
         patchMicroAppRoute(route, getMicroAppRouteComponent, {
           base,
           masterHistoryType,
@@ -104,7 +106,7 @@ function useLegacyModifyRoutesWithRegistrableMode(
     return null;
   };
 
-  return routes.map(route => {
+  return routes.map((route) => {
     if (route.path === '/' && route.routes && route.routes.length) {
       apps.forEach(({ history: slaveHistory = masterHistoryType, base }) => {
         if (!base) {
@@ -115,7 +117,7 @@ function useLegacyModifyRoutesWithRegistrableMode(
         if (slaveHistory === masterHistoryType) {
           const baseConfig = toArray(base);
 
-          baseConfig.forEach(basePath => {
+          baseConfig.forEach((basePath) => {
             const routeWithPrefix = findRouteWithPrefix(routes, basePath);
 
             // 应用没有自己配置过 basePath 相关路由，则自动加入 mock 的路由